fix(tag): post review changes
